+++ packages/at_commons/lib/src/at_constants.dart
@@ -28,6 +28,8 @@ class AtConstants {
   static const String ttr = 'ttr';
   static const String ttlNotification = 'ttln';
   static const String fromCommitSequence = 'from_commit_seq';
+  static const String syncLimit = 'limit';
+  static const String skipDeletesUntil = 'skipDeletesUntil';
   static const String atValueReference = 'atsign://';
   static const String regex = 'regex';
   static const String configNew = 'configNew';